Refactored to combine adapters
